+++ Fun/Rolling.lua
@@ -1,8 +1,5 @@
 local _, addon = ...
 
--- Plays a random rolling sound when the player uses Roll or Chi Torpedo (Monk).
--- Add spell IDs to MI_ROLL_IDS to extend coverage.
-
 local SOUND_DIR = "Interface\\AddOns\\MysteriousQoL\\Sounds\\Rolling\\"
 
 local SOUNDS = {
@@ -11,15 +8,14 @@ local SOUNDS = {
 }
 
 -- Add any roll/dash spell ID here.
-local MI_ROLL_IDS = {
+local SPELL_IDS = {
     [109132] = true,  -- Roll         (Monk)
     [115008] = true,  -- Chi Torpedo  (Monk talent)
 }
 
-local f = CreateFrame("Frame")
-f:RegisterUnitEvent("UNIT_SPELLCAST_SUCCEEDED", "player")
-f:SetScript("OnEvent", function(_, _, _, _, spellID)
-    if addon.db.fun_rolling_enabled and MI_ROLL_IDS[spellID] then
-        PlaySoundFile(SOUNDS[math.random(#SOUNDS)], addon.db.fun_rolling_channel)
-    end
-end)
+addon.MI_CreateSpellCastSound({
+    dbKey      = "fun_rolling_enabled",
+    channelKey = "fun_rolling_channel",
+    spellIDs   = SPELL_IDS,
+    sounds     = SOUNDS,
+})